+++ vlib/v/gen/fn.v
@@ -47,7 +47,7 @@ fn (mut g Gen) gen_fn_decl(it ast.FnDecl, skip bool) {
 	}
 	//
 	mut name := it.name
-	if name[0] in [`+`, `-`, `*`, `/`, `%`] {
+	if name[0] in [`+`, `-`, `*`, `/`, `%`, `<`, `>`] {
 		name = util.replace_op(name)
 	}
 	if it.is_method {
@@ -424,6 +424,11 @@ fn (mut g Gen) method_call(node ast.CallExpr) {
 			}
 		}
 	}
+	if node.generic_type != table.void_type && node.generic_type != 0 {
+		// Using _T_ to differentiate between get<string> and get_string
+		// `foo<int>()` => `foo_T_int()`
+		name += '_T_' + g.typ(node.generic_type)
+	}
 	// TODO2
 	// g.generate_tmp_autofree_arg_vars(node, name)
 	//
@@ -443,7 +448,8 @@ fn (mut g Gen) method_call(node ast.CallExpr) {
 		if !is_range_slice {
 			g.write('&')
 		}
-	} else if !node.receiver_type.is_ptr() && node.left_type.is_ptr() && node.name != 'str' {
+	} else if !node.receiver_type.is_ptr() && node.left_type.is_ptr() && node.name != 'str' &&
+		node.from_embed_type == 0 {
 		g.write('/*rec*/*')
 	}
 	if node.free_receiver && !g.inside_lambda && !g.is_builtin_mod {